+++ src/lib/index.ts
@@ -1,6 +1,5 @@
 export { createUndoRedoSignal } from "./travel";
 export type {
-  Setter,
   UndoRedoSignal,
   UndoRedoSignalOptions,
   OnRedoCallback,