+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.0
+
+- Now you don't need to always pass an initial value to the signal
+- `reactiveHistoryGenerator` => `createHistoryIterator`
+- `Setter` type is not exported anymore. It's the same as in solid-js
+
 # 1.1.0
 
 - Added support for onUndo and onRedo callbacks